+++ pkg/connect/termios_darwin.go
@@ -1,0 +1,13 @@
+//go:build darwin
+
+package connect
+
+import "golang.org/x/sys/unix"
+
+func ioctlReadTermios() uint {
+	return unix.TIOCGETA
+}
+
+func ioctlWriteTermios() uint {
+	return unix.TIOCSETA
+}